Setting variables for binding buffers

+++ Tutorial/BaseObject.h
@@ -25,7 +25,7 @@ public:
 
 	virtual void Init() = 0;
 	virtual float* GetVertexBuffer() = 0;
-	virtual float* GetIndexBuffer() = 0;
+	virtual int* GetIndexBuffer() = 0;
 	virtual float* GetColorBuffer() = 0;
 
 	virtual void Draw() = 0;
@@ -37,7 +37,7 @@ protected:
 	float rotation[3];
 	float scale[3];
 
-	float *ib;	//index buffer
+	int *ib;	//index buffer
 	float *vb;	//vertex buffer
 	float *nb;	//normal buffer
 	float *cb;	//color buffer
