+++ Tutorial/ObjectTriangle.h
@@ -13,7 +13,7 @@ public:
 
 	virtual void Init();
 	virtual float* GetVertexBuffer();
-	virtual float* GetIndexBuffer();
+	virtual int* GetIndexBuffer();
 	virtual float* GetColorBuffer();
 
 	virtual void Draw();
